[ias-verify] extract ephemeral_key.rs
